refactor: transformer
