+++ config/model/yaak/control_transformer/raw.yaml
@@ -302,7 +302,7 @@ episode_builder:
     _target_: rmind.components.mask.CausalAttentionMaskBuilder
 
 encoder:
-  _target_: rmind.components.llm.TransformerEncoder
+  _target_: rmind.components.transformer.TransformerEncoder
   dim_model: ${encoder_embedding_dim}
   num_layers: ${num_layers}
   num_heads: ${num_heads}
@@ -431,9 +431,9 @@ objectives:
               bias: False
           foresight:
             cam_front_left:
-              _target_: rmind.components.llm.CrossAttentionDecoderHead
+              _target_: rmind.components.transformer.CrossAttentionDecoderHead
               decoder:
-                _target_: rmind.components.llm.CrossAttentionDecoder
+                _target_: rmind.components.transformer.CrossAttentionDecoder
                 dim_model: ${image_embedding_dim}
                 num_layers: 2
                 num_heads: 4
