+++ config/model/yaak/control_transformer/raw_export.yaml
@@ -217,7 +217,7 @@ episode_builder:
     _target_: rmind.components.mask.CausalAttentionMaskBuilder
 
 encoder:
-  _target_: rmind.components.llm.TransformerEncoder
+  _target_: rmind.components.transformer.TransformerEncoder
   dim_model: ${encoder_embedding_dim}
   num_layers: ${num_layers}
   num_heads: ${num_heads}
