+++ src/rmind/components/transformer/decoder.py
@@ -1,0 +1,146 @@
+from typing import Self, final, override
+
+from pydantic import BaseModel, ConfigDict, model_validator, validate_call
+from torch import Tensor, nn
+
+from rmind.components.transformer._runtime import freeze_module, run_layer_stack
+from rmind.components.transformer.feed_forward import MLPGLU
+
+
+class CrossAttentionDecoderBlock(nn.Module):
+    @validate_call
+    def __init__(  # noqa: PLR0913, PLR0917
+        self,
+        embedding_dim: int,
+        num_heads: int,
+        attn_dropout: float = 0.1,
+        resid_dropout: float = 0.1,
+        mlp_dropout: float = 0.1,
+        hidden_layer_multiplier: int = 1,
+    ) -> None:
+        super().__init__()
+
+        self.cross_attn_norm = nn.LayerNorm(embedding_dim)
+        self.cross_attn = nn.MultiheadAttention(
+            embed_dim=embedding_dim,
+            num_heads=num_heads,
+            dropout=attn_dropout,
+            batch_first=True,
+        )
+        self.cross_attn_resid_drop = nn.Dropout(resid_dropout, inplace=False)
+
+        self.self_attn_norm = nn.LayerNorm(embedding_dim)
+        self.self_attn = nn.MultiheadAttention(
+            embed_dim=embedding_dim,
+            num_heads=num_heads,
+            dropout=attn_dropout,
+            batch_first=True,
+        )
+        self.self_attn_resid_drop = nn.Dropout(resid_dropout, inplace=False)
+
+        self.mlp_norm = nn.LayerNorm(embedding_dim)
+        self.mlp = MLPGLU(
+            dim_model=embedding_dim,
+            dropout=mlp_dropout,
+            hidden_layer_multiplier=hidden_layer_multiplier,
+        )
+
+    @override
+    def forward(self, x: Tensor, context: Tensor) -> Tensor:
+        residual = x
+        x_norm = self.cross_attn_norm(x)
+        cross_attn_out, _ = self.cross_attn(
+            query=x_norm, key=context, value=context, need_weights=False
+        )
+        x = residual + self.cross_attn_resid_drop(cross_attn_out)
+
+        residual = x
+        x_norm = self.self_attn_norm(x)
+        self_attn_out, _ = self.self_attn(
+            query=x_norm, key=x_norm, value=x_norm, need_weights=False
+        )
+        x = residual + self.self_attn_resid_drop(self_attn_out)
+
+        residual = x
+        mlp_out = self.mlp(self.mlp_norm(x))
+        return residual + mlp_out
+
+
+class CrossAttentionDecoder(nn.Module):
+    def __init__(  # noqa: PLR0913, PLR0917
+        self,
+        dim_model: int,
+        num_layers: int,
+        num_heads: int,
+        attn_dropout: float = 0.1,
+        resid_dropout: float = 0.1,
+        mlp_dropout: float = 0.1,
+        hidden_layer_multiplier: int = 1,
+        freeze: bool | None = None,  # noqa: FBT001
+    ) -> None:
+        super().__init__()
+        self.layers = nn.ModuleList([
+            CrossAttentionDecoderBlock(
+                embedding_dim=dim_model,
+                num_heads=num_heads,
+                attn_dropout=attn_dropout,
+                mlp_dropout=mlp_dropout,
+                resid_dropout=resid_dropout,
+                hidden_layer_multiplier=hidden_layer_multiplier,
+            )
+            for _ in range(num_layers)
+        ])
+        self.layer_norm = nn.LayerNorm(dim_model)
+        freeze_module(self, freeze=freeze)
+
+    @override
+    def forward(self, x: Tensor, context: Tensor) -> Tensor:
+        x = run_layer_stack(self.layers, x, context, training=self.training)
+        return self.layer_norm(x)
+
+
+@final
+class CrossAttentionDecoderHead(nn.Module):
+    class Input(BaseModel):
+        model_config = ConfigDict(arbitrary_types_allowed=True, extra="forbid")
+
+        query: Tensor
+        context: Tensor
+
+        @model_validator(mode="after")
+        def _validate_shapes(self) -> Self:
+            if self.query.ndim != self.context.ndim or self.query.ndim not in {3, 4}:
+                msg = (
+                    "query/context must both be 3D or 4D with matching ndim, "
+                    f"got query={self.query.ndim}D, context={self.context.ndim}D"
+                )
+                raise ValueError(msg)
+            return self
+
+    def __init__(
+        self, decoder: CrossAttentionDecoder, output_projection: nn.Linear
+    ) -> None:
+        super().__init__()
+        self.decoder = decoder
+        self.output_projection = output_projection
+
+    @validate_call
+    @override
+    def forward(self, input: Input) -> Tensor:
+        query = input.query
+        context = input.context
+
+        if query.ndim == 4:  # noqa: PLR2004
+            b, t, sq, d = query.shape
+            _, _, sc, _ = context.shape
+
+            query_flat = query.reshape(b * t, sq, d)
+            context_flat = context.reshape(b * t, sc, d)
+
+            decoded = self.decoder(query_flat, context_flat)
+            output = self.output_projection(decoded)
+
+            return output.reshape(b, t, sq, d)
+
+        decoded = self.decoder(query, context)
+        return self.output_projection(decoded)
