+++ src/rmind/components/transformer/feed_forward.py
@@ -1,0 +1,28 @@
+from typing import final, override
+
+from torch import Tensor, nn
+
+
+@final
+class MLPGLU(nn.Module):
+    def __init__(
+        self,
+        dim_model: int,
+        dropout: float,
+        hidden_layer_multiplier: int,
+        bias: bool = True,  # noqa: FBT001, FBT002
+    ) -> None:
+        super().__init__()
+        dim_mlp = hidden_layer_multiplier * dim_model
+        self.l1 = nn.Linear(in_features=dim_model, out_features=dim_mlp * 2, bias=bias)
+        self.a1 = nn.GELU()
+        self.d1 = nn.Dropout(dropout)
+        self.l2 = nn.Linear(in_features=dim_mlp, out_features=dim_model, bias=bias)
+
+    @override
+    def forward(self, input: Tensor) -> Tensor:
+        # FFN_GEGLU eq. 6, https://arxiv.org/pdf/2002.05202v1.pdf
+        x = self.l1(input)
+        xw, xv = x.chunk(2, dim=-1)
+        geglu = self.a1(xw) * xv
+        return self.l2(self.d1(geglu))
